requirements.txt
# Discord library with voice support (Python 3.13 compatible)
discord.py[voice]>=2.6.0
PyNaCl==1.5.0

# Environment management
python-dotenv==1.0.0

# Audio processing
ffmpeg-python==0.2.0
pyaudiowpatch>=0.2.12.7  # WASAPI loopback for system audio capture (Windows)

# HTTP requests for Icecast stream
requests==2.31.0
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1
